+++ trunk/OgreMain/src/OgrePanelOverlayElement.cpp
@@ -263,7 +263,8 @@ namespace Ogre {
                 // remove extras
                 for (size_t i = mNumTexCoordsInBuffer; i > numLayers; --i)
                 {
-                    decl->removeElement(VES_TEXTURE_COORDINATES, i);
+                    decl->removeElement(VES_TEXTURE_COORDINATES, 
+						static_cast<unsigned short>(i));
                 }
             }
             else if (mNumTexCoordsInBuffer < numLayers)
@@ -273,7 +274,8 @@ namespace Ogre {
                 for (size_t i = mNumTexCoordsInBuffer; i < numLayers; ++i)
                 {
                     decl->addElement(TEXCOORD_BINDING,
-                        offset, VET_FLOAT2, VES_TEXTURE_COORDINATES, i);
+                        offset, VET_FLOAT2, VES_TEXTURE_COORDINATES, 
+						static_cast<unsigned short>(i));
                     offset += VertexElement::getTypeSize(VET_FLOAT2);
 
                 }
